% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCollapse.R
\name{collapseGenes}
\alias{collapseGenes}
\title{Collapse Genes}
\usage{
collapseGenes(ge, probes)
}
\arguments{
\item{ge}{matrix, rows are probes, columns are samples}

\item{probes}{df, matrix or named vector}
}
\value{
collapsed gene expression matrix with probes replaced with corresponding genes
}
\description{
Collapses given dataset so every gene is presented by the highest expressed probe
}
