% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProportions.R
\name{dotPlotPropotions}
\alias{dotPlotPropotions}
\title{Proportions dot plot}
\usage{
dotPlotPropotions(predicted, actual, guess = FALSE, main = NULL,
  showR2 = FALSE)
}
\arguments{
\item{predicted}{matrix of predicted proportions}

\item{actual}{matrix of actual proportions}

\item{guess}{if True will function will try to guess how to reorder rows of predicted proportions to match rows of actual proportions}

\item{main}{plot title}

\item{showR2}{calculate and show R squared statistics}
}
\description{
Proportions dot plot
}
