% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateNoise.R
\name{estimateNoise}
\alias{estimateNoise}
\title{Noise estimation}
\usage{
estimateNoise(Y, noiseType = "additive", verbose = FALSE)
}
\arguments{
\item{Y}{normalized gene expression data matrix, columns are genes and rows are samples}

\item{noiseType}{character, describing noise type. Two possible values are "additive" and "possion"}

\item{verbose}{logical, default value is FALSE}
}
\value{
list with two elements, w -- estimated noise and Rw estimated noise correlation matrix
}
\description{
Estimates noise using multiple regression approach. Implements method described in
J. M. Bioucas-Dias and J. M. P. Nascimento, "Hyperspectral Subspace Identification," in IEEE Transactions on Geoscience and Remote Sensing, vol. 46, no. 8, pp. 2435-2445, Aug. 2008.
}
\details{
Based on MATLAB original code from http://www.lx.it.pt/~bioucas/code.htm
}
