% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa.R
\name{fastDSA}
\alias{fastDSA}
\title{Fast DSA algorithm implementation}
\usage{
fastDSA(dataset, genes)
}
\arguments{
\item{dataset}{gene expression matrix}

\item{genes, }{list with putative signatures for DSA algorithm}
}
\value{
deconvolution results, list with H and W matrices
}
\description{
Runs DSA implementation using .fccnls for solving least-squares with multiple right-hand-sides
}
