% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{generateMixedData}
\alias{generateMixedData}
\title{Generate mixed data with or without noise}
\usage{
generateMixedData(samples = 40, genes = 12000, cellTypes = 3,
  bias = NULL, spearmanThreshold = 0.5, pureGenes = 0,
  noiseDeviation = 0, sampleLogMean = 4, sampleLogSd = 3,
  removeAngles = F, cutCoef = 0.85, removeBorders = F,
  borderShift = 0.25)
}
\arguments{
\item{samples}{number of samples}

\item{genes}{number of genes}

\item{cellTypes}{number of cell types}

\item{bias}{if not null bias should be a number in between 0 and 1, one of cell types presented in mix will be more abundunt and one cell type will be less abundunt}

\item{pureGenes}{number of genes which are not noisy}

\item{noiseDeviation}{standart deviation of normally distributed noise value}

\item{removeAngles}{if TRUE there will be no signature genes in the mix (simplex corners will be removed)}

\item{removeBorders}{if TRUE where will be no genes close to simplex border}

\item{borderShift}{number in between 0 and 1, every value in basis is guaranteed to be at least borderShift}

\item{cutCoeff}{corners are removed as points that lie outside of cellTypes-dimensional sphere with radius of cutCoef}
}
\description{
Generates mixed data with or without noise under assumption of linear model
}
