% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectiveProjection.R
\name{getProjectiveProjection}
\alias{getProjectiveProjection}
\title{Get projective projection}
\usage{
getProjectiveProjection(Y, p, spherize = F)
}
\arguments{
\item{Y}{High dimensional data to project}

\item{p}{Dimensionality to project to}

\item{spherize}{Spherize dataset or not}
}
\value{
projection
}
\description{
Get projective projection
}
