% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProportions.R
\name{guessOrder}
\alias{guessOrder}
\title{guess the order}
\usage{
guessOrder(predicted, actual)
}
\arguments{
\item{predicted}{predicted propotions}

\item{actual}{actual proportions}
}
\value{
numeric, correct order of predicted proportions
}
\description{
Function tries to guess ordering for rows of predicted proportions to match rows of actual proportions
}
