% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hinge.R
\name{hinge}
\alias{hinge}
\title{Hinge function}
\usage{
hinge(Y)
}
\arguments{
\item{Y}{matrix describing genes (possibly in reduced space)}
}
\value{
matrix -Y with negative values replaced with zeroes
}
\description{
Hinge function
}
