% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hysime.R
\name{hysime}
\alias{hysime}
\title{HySime}
\usage{
hysime(Y, W, Rn, verbose = FALSE)
}
\arguments{
\item{Y}{gene expression matrix, columns are genes, rows are samples}

\item{W}{estimated noise matrix}

\item{Rn}{estimated noise correlation matrix}

\item{verbose}{verbosity, default valie is FALSE}
}
\value{
list
}
\description{
Evalutes number of cell types presented in mixture using (HySime) hyperspectral signal identification by minimum error
}
\details{
Original paper is J. M. Bioucas-Dias and J. M. P. Nascimento, "Hyperspectral Subspace Identification," in IEEE Transactions on Geoscience and Remote Sensing, vol. 46, no. 8, pp. 2435-2445, Aug. 2008.

Based on MATLAB original code from http://www.lx.it.pt/~bioucas/code.htm
}
