% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{linearizeDataset}
\alias{linearizeDataset}
\title{linearizeDataset}
\usage{
linearizeDataset(ge)
}
\arguments{
\item{ge}{gene expression matrix}
}
\value{
gene expression matrix in linear scale
}
\description{
linearizeDataset
}
