% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linseedObject.R
\docType{class}
\name{LinseedObject}
\alias{LinseedObject}
\title{Linseed Object}
\format{\code{\link{R6Class}} object.}
\usage{
LinseedObject
}
\value{
Object of \code{\link{R6Class}} -- an interface to work with gene expression data.
}
\description{
The Linseed Object Class.
}
\details{
Provides an interface to perform 
collinear network construction,
linear subspace identification, 
simplex corner identification and gene expression deconvolution
}
\section{Fields}{

\describe{
\item{\code{exp}}{List of two elements raw and normalized gene expression dataset}

\item{\code{name}}{Character, optional, dataset name}

\item{\code{cellTypeNumber}}{Identified cell type number, required for projection, 
corner detection and deconvolution}

\item{\code{projection}}{Projection of genes into space lower-dimensionality (presumably simplex)}

\item{\code{endpoints}}{Simplex corners (in normalized, non-reduced space)}

\item{\code{endpointsProjection}}{Simplex corners (in reduced space)}

\item{\code{distances}}{Stores distances for every gene to each corner in reduced space}

\item{\code{markers}}{List that stores signatures genes for deconvolution, can be set manually or can be obtained by \code{selectGenes(k)}}

\item{\code{signatures}}{Deconvolution signature matrix}

\item{\code{proportions}}{Deconvolution proportion matrix}

\item{\code{pairwise}}{Calculated pairwise collinearity measure}
}}

\section{Methods}{

\describe{

}
}

\examples{
LinseedObject$new("GSE19830", samples=10:42, topGenes=10000)


}
\keyword{datasets}
