% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProportions.R
\name{plotProportions}
\alias{plotProportions}
\title{Draw a plot of estimated proportions}
\usage{
plotProportions(..., pnames = NULL, point_size = 2, line_size = 1)
}
\arguments{
\item{...}{matricies, data frames, NMF objects of estimated proportions or paths to file}

\item{pnames}{experiment titles}
}
\value{
ggplot object
}
\description{
Draws a plot of estimated proprotions
If ggplot2 and reshape2 are installed will use them and return ggplot object
Otherwise will use standart R functions
}
\examples{
set.seed(31)
data('datasetLiverBrainLung')
mixed <- datasetLiverBrainLung[, 10:42]
preprocessed <- preprocessDataset(mixed, k=5) # 5 clusters
accuracy <- clusdecAccuracy(preprocessed, 3) # assuming 3 cell types
results <- chooseBest(preprocessed, accuracy) # choose best combination of clusters as putative sigantures

plotProportions(results$H) # plotting proportions

data('proportionsLiverBrainLung') # or we can compare with actual mixed proportions
mixedProportions <- proportionsLiverBrainLung[, 10:42]
plotProportions(results$H, mixedProportions, pnames=c('ClusDec estimated', 'Actual'))

}
