% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplexUtils.R
\name{plotRotation}
\alias{plotRotation}
\title{Rotation plotter}
\usage{
plotRotation(reduced, dims, step = 5, colors = NULL,
  fileName = "rotation.gif")
}
\arguments{
\item{reduced}{}

\item{dims}{}

\item{step}{}

\item{colors}{}

\item{fileName}{}
}
\description{
Useful rotation plotter for high-dimensional simplexes
}
