% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocessDataset}
\alias{preprocessDataset}
\title{Preprocess Dataset}
\usage{
preprocessDataset(dataset, annotation = NULL, geneSymbol = "Gene Symbol",
  samples = NULL, topGenes = 10000, topVar = FALSE, ...)
}
\arguments{
\item{dataset}{matrix, data.frame, path to file or GSE accession with expression data}

\item{annotation}{dataframe, matrix, named vector with annotation to probes}

\item{geneSymbol}{column from annotation to collapse the genes, deafult value is 'Gene Symbol'}

\item{samples}{character vector of samples. If column were not in samples, it would be excluded from analysis.
Default value is NULL, which takes every sample from dataset}

\item{topGenes}{integer How many genes include in analysis. We suppose to include only expressed genes. Default value is 10000}
}
\value{
clustered dataset, matrix, first column identifies cluster of the row
}
\description{
Preprocesses given dataset. Preprocessing consists of 3 major steps:
1) If needed, probes corresponding to the same genes are collapsed, only most expressed probe is taken for further analysis.
   It's common technique in microarray data analysis.
2) If needed, only highly expressed genes are taken for further analysis. (Say hello to noize reduction)
3) All genes are clustered with Kmeans using cosine simillarity as distance.
}
\examples{
data('datasetLiverBrainLung')
prep <- preprocessDataset(datasetLiverBrainLung)
prep <- preprocessDataset(datasetLiverBrainLung, k=5) # 5 clusters
prep <- preprocessDataset(datasetLiverBrainLung, topGenes=6000) # leave only top 6k genes

}
