% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocessGSE}
\alias{preprocessGSE}
\title{Preprocess GSE Dataset}
\usage{
preprocessGSE(geoAccesion, annotate = TRUE, normalize = TRUE, ...)
}
\arguments{
\item{geoAccesion}{e.g 'GSE19830'}

\item{annotate}{annotate with feature data from provided geo platform}

\item{normalize}{quantile normalize GEO dataset}

\item{...}{arguments further passed to preprocessDataset}
}
\value{
clustered dataset, matrix, first column identifies cluster of the row
}
\description{
Downloads GSE dataset by GEO accession and performs preprocessing
}
