% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectiveProjection.R
\name{projectiveProjection}
\alias{projectiveProjection}
\title{Projective projection}
\usage{
projectiveProjection(Y, p, spherize = F)
}
\arguments{
\item{Y}{High dimensional data to project}

\item{p}{Dimensionality to project to}

\item{spherize}{Spherize dataset or not}
}
\value{
matrix with coordinates of projected points.
}
\description{
Projective projection
}
