% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsa.R
\name{pureDsa}
\alias{pureDsa}
\title{DSA algorithm implementation for pure points}
\usage{
pureDsa(dataset, pure)
}
\arguments{
\item{dataset}{gene expression matrix}

\item{pure}{matrix contains expression of signature genes}
}
\value{
deconvolution results, list with H and W matrices
}
\description{
Runs DSA implementation using fcnnls_c for solving least-squares with multiple right-hand-sides
}
