% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateData.R
\name{sampleFromSimplexUniformly}
\alias{sampleFromSimplexUniformly}
\title{Generation of points uniformly distributed on k-dimensional standard simplex}
\usage{
sampleFromSimplexUniformly(n, k = 3, M = 1e+05)
}
\arguments{
\item{n}{number of poitns}

\item{k}{dimensionality}

\item{M}{grid size}
}
\value{
matrix where columns are points
}
\description{
Generation of points uniformly distributed on k-dimensional standard simplex
}
