% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sisal.R
\name{sisal}
\alias{sisal}
\title{SISAL algorithm}
\usage{
sisal(Y, p, iters = 80, tau = 1, mu = p * 1000/ncol(Y), spherize = F,
  tol = 0.01, m0 = NULL, verbose = F, returnPlot = T, nonNeg = F)
}
\arguments{
\item{Y}{gene expression matrix}

\item{p}{number of endpoints}

\item{iters}{number of iterations to perform}

\item{tau}{noise points penalty coefficient}

\item{mu}{regularization}

\item{spherize}{spherize or not}

\item{tol}{numeric tolerance}

\item{m0}{starting points for SISAL algorithm, default points are getting from VCA}

\item{verbose}{verbosity}

\item{returnPlot}{logical, is it needed to return dataframe or not}
}
\description{
Sisal alogorithm for simplex endpoints identification
}
\details{
Implementation of method
}
