% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vca.R
\name{vca}
\alias{vca}
\title{Title}
\usage{
vca(R, p, SNR = NULL, verbose = F)
}
\arguments{
\item{R}{matrix describing points (possibly lying in a simplex) in high dimensional space}

\item{p}{number endpoints to find}

\item{SNR}{signal to noise ratio, NULL by default}

\item{verbose}{verbosity, deafult value is FALSE}
}
\value{
matrix of columns from R which are considered to be endpoints
}
\description{
Title
}
