% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateNoise.R
\name{estimateAdditiveNoise}
\alias{estimateAdditiveNoise}
\title{Estimate additive noise}
\usage{
estimateAdditiveNoise(Y, verbose)
}
\arguments{
\item{Y}{normalized gene expression matrix}

\item{verbose}{verbosity}
}
\value{

}
\description{
Additive noise estimation subroutine
}
