% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProportions.R
\name{plotProportions}
\alias{plotProportions}
\title{Draw a plot of estimated proportions}
\usage{
plotProportions(..., pnames = NULL, point_size = 2, line_size = 1)
}
\arguments{
\item{...}{matricies, data frames, NMF objects of estimated proportions or paths to file}

\item{pnames}{experiment titles}

\item{point_size}{point size for plot}

\item{line_size}{line size for plot}
}
\value{
ggplot object
}
\description{
Draws a plot of estimated proprotions
If ggplot2 and reshape2 are installed will use them and return ggplot object
Otherwise will use standart R functions
}
