% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectiveProjection.R
\name{projectiveProjection}
\alias{projectiveProjection}
\title{Projective projection}
\usage{
projectiveProjection(Y, p, spherize = F)
}
\arguments{
\item{Y}{}

\item{p}{}

\item{spherize}{}
}
\description{
Projective projection
}
